面试项目集/01_CV_包裹检测系统/requirements.txt
# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# 图像处理
opencv-python>=4.8.0
Pillow>=10.0.0

# 数据处理
numpy>=1.24.0
pandas>=2.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 工具
tqdm>=4.65.0
pyyaml>=6.0
scikit-learn>=1.3.0
